+++ lib/src/widgets/turnable_pdf.dart
@@ -3,7 +3,6 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:pdfrx/pdfrx.dart';
 import 'package:turnable_page/src/widgets/paper_widget.dart';